DAG-570: Gjøre om inngang siden til første siden av appen
