+++ src/components/exit-soknad/DeleteSoknadModal.tsx
@@ -75,7 +75,7 @@ export function DeleteSoknadModal({ isOpen, handleClose }: IProps) {
                 <Button as="a">{getAppText("soknad-slettet.modal.knapp.lukk")}</Button>
               </Link>
 
-              <Link href="/" passHref>
+              <Link href="/start-soknad" passHref>
                 <Button as="a" variant="tertiary">
                   {getAppText("soknad-slettet.modal.knapp.start-ny-soknad")}
                 </Button>
